% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CSMF2}
\alias{CSMF2}
\title{Create CSMF plot with Undetermined for InterVA4 Output}
\usage{
CSMF2(x, top)
}
\arguments{
\item{x}{An fitted object returned from openVA::codeVA()}

\item{top}{An integer specifying the number of top causes included
in the results.}
}
\value{

}
\description{
Create CSMF plot with Undetermined for InterVA4 Output
}
